+++ SyncKusto.sln
@@ -9,6 +9,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SyncKusto.Tests", "SyncKust
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8EC462FD-D22E-90A8-E5CE-7E832BA40C5D}"
 	ProjectSection(SolutionItems) = preProject
+		Additional-Refactoring.md = Additional-Refactoring.md
 		REFACTORING_PLAN.md = REFACTORING_PLAN.md
 	EndProjectSection
 EndProject
